requirements.txt
langchain
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf
mysql-connector-python
pymysql
sentence-transformers
SQLAlchemy
chromadb

